requirements.txt
aio-pika==6.7.1
aiormq==3.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
fastapi==0.68.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
multidict==6.0.4
pamqp==2.3.0
psycopg2==2.9.9
pydantic==1.10.13
pydantic_core==2.14.3
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.14.2
typing_extensions==4.8.0
uvicorn==0.24.0.post1
vine==5.1.0
yarl==1.9.2
